TASK-26265 centralize tomcat version in maven-depmgt-pom
