+++ test/servers/tomcat8/pom.xml
@@ -22,9 +22,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
       <version>6.2.x-SNAPSHOT</version>
    </parent>
    <modelVersion>4.0.0</modelVersion>
-   <artifactId>wci-test-tomcat8</artifactId>
+   <artifactId>wci-test-tomcat</artifactId>
    <packaging>jar</packaging>
-   <name>GateIn - WCI test Tomcat8</name>
+   <name>GateIn - WCI test Tomcat</name>
 
   <properties>
     <exo.test.coverage.ratio>0.75</exo.test.coverage.ratio>
@@ -45,7 +45,7 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
     <!-- -->
     <dependency>
       <groupId>org.exoplatform.gatein.wci</groupId>
-      <artifactId>wci-tomcat8</artifactId>
+      <artifactId>wci-tomcat</artifactId>
     </dependency>
 
     <!-- Arquillian container -->
