+++ test/servers/tomcat8/src/test/resources/org/gatein/wci/test/tomcat/crosscontext/web.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0"?>
+<!--
+This file is part of the Meeds project (https://meeds.io/).
+Copyright (C) 2020 Meeds Association
+contact@meeds.io
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU Lesser General Public
+License as published by the Free Software Foundation; either
+version 3 of the License, or (at your option) any later version.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+Lesser General Public License for more details.
+You should have received a copy of the GNU Lesser General Public License
+along with this program; if not, write to the Free Software Foundation,
+Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+-->
+<web-app xmlns="http://java.sun.com/xml/ns/javaee"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
+         version="3.0">
+
+  <servlet>
+     <servlet-name>ContainerServlet</servlet-name>
+     <servlet-class>org.gatein.wci.tomcat.TomcatContainerServlet</servlet-class>
+     <load-on-startup>0</load-on-startup>
+  </servlet>
+  <servlet>
+    <servlet-name>CrossContextServlet</servlet-name>
+    <servlet-class>org.gatein.wci.test.crosscontext.CrossContextServlet</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>CrossContextServlet</servlet-name>
+    <url-pattern>/</url-pattern>
+  </servlet-mapping>
+
+  <login-config>
+    <auth-method>BASIC</auth-method>
+    <realm-name>MyUserDatabase</realm-name>
+  </login-config>
+
+  <security-role>
+    <role-name>myrole</role-name>
+  </security-role>
+
+  <session-config>
+    <tracking-mode>URL</tracking-mode>
+  </session-config>
+
+</web-app>
