+++ test/servers/tomcat8/src/test/resources/org/gatein/wci/test/tomcat/deployment/context.xml
