+++ tomcat/pom.xml
@@ -23,10 +23,18 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
    </parent>
    <modelVersion>4.0.0</modelVersion>
    <artifactId>wci-tomcat</artifactId>
-   <packaging>pom</packaging>
+   <packaging>jar</packaging>
    <name>GateIn - WCI Tomcat compatibility component</name>
 
-   <modules>
-      <module>tomcat8</module>
-   </modules>
+   <dependencies>
+      <dependency>
+         <groupId>org.exoplatform.gatein.wci</groupId>
+         <artifactId>wci-wci</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>org.apache.tomcat</groupId>
+         <artifactId>tomcat-catalina</artifactId>
+         <scope>provided</scope>
+      </dependency>
+   </dependencies>
 </project>
