+++ tomcat/src/main/java/org/gatein/wci/command/TomcatCommandDispatcher.java
